add: pytest plugin for noct mocks
