+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     hooks:
       - id: isort
         name: isort (python)
-        args: ["--profile", "black", "--filter-files"]
+        args: ["--profile", "black", "--filter-files", "--project", "virga"]
   - repo: https://github.com/psf/black
     rev: 22.3.0
     hooks:
